// src/app/modules/motiapi/model/ricercaIntervento.ts
/*
* SPDX-FileCopyrightText: Copyright 2020 - 2021 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Moti
 * API per il backend della suite di Moti.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { FonteDati } from './fonteDati';
import { Luogo } from './luogo';
import { Tipologia } from './tipologia';
import { Utente } from './utente';


export interface RicercaIntervento { 
    codIntervento?: string;
    fonteDati?: FonteDati;
    luogos?: Array<Luogo>;
    referentes?: Array<Utente>;
    tipologia?: Tipologia;
    titolo?: string;
}